Add Driver Management System
- Create Driver model with relationships to Transport
- Add drivers table migration with unique constraints
- Implement DriverResource with Filament v4 structure
- Configure DriverForm with sections for basic info, license, and photos
- Set up DriversTable with searchable columns and image display
- Create DriverSeeder with 10 sample drivers
- Add Russian localization for all labels
- Include unique constraints for phone, email, and license number
- Support for profile and license image uploads
- Ready for integration with Transport system

+++ database/migrations/2025_09_28_114955_create_companies_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('phone')->nullable();
             $table->string('email')->nullable();
             $table->integer('inn')->nullable();
-            $table->integer('account_number')->nullable();
+            $table->bigInteger('account_number')->nullable();
             $table->string('bank_name')->nullable();
             $table->integer('bank_mfo')->nullable();
             $table->string('director_name')->nullable();
